pyproject.toml
[project]
name = "gdg-2026"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.73",
    "streamlit>=1.40",
    "plotly>=5.24",
]
